refactor:     Moved API_KEY to keystore.properties, so that it was not published to a public repo
