+++ Week8/MovieViewer/app/src/main/AndroidManifest.xml
@@ -13,14 +13,14 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".activities.LoginActivity">
+        <activity android:name=".ui.LoginActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".activities.DetailActivity" />
-        <activity android:name=".activities.MainActivity" />
+        <activity android:name=".ui.DetailActivity" />
+        <activity android:name=".ui.MainActivity" />
     </application>
 
 </manifest>
