+++ Week8/MovieViewer/app/src/main/java/com/wmw/movieviewer/ViewExtensions.kt
@@ -7,6 +7,6 @@ fun View.onClick(onClickAction: () -> Unit) {
     setOnClickListener { onClickAction() }
 }
 
-fun View.longPress(onLongPressAction: () -> Boolean) {
-    setOnLongClickListener { onLongPressAction() }
-}
+//fun View.longPress(onLongPressAction: () -> Boolean) {
+//    setOnLongClickListener { onLongPressAction() }
+//}
