+++ Week8/MovieViewer/app/src/main/java/com/wmw/movieviewer/model/MovieDao.kt
@@ -5,10 +5,10 @@ import androidx.room.*
 @Dao
 interface MovieDao {
     @Insert(onConflict = OnConflictStrategy.REPLACE)
-    suspend fun insertMovies(movies: List<Movie>)
+    suspend fun insert(movies: List<Movie>)
 
     @Insert(onConflict = OnConflictStrategy.REPLACE)
-    suspend fun insertMovie(movie: Movie)
+    suspend fun insert(movie: Movie)
 
     @Delete
     suspend fun deleteMovies(vararg movie: Movie)
@@ -17,7 +17,7 @@ interface MovieDao {
     suspend fun deleteMovieById(movieId: String)
 
     @Query("SELECT * FROM movies_table ORDER BY title ASC")
-    suspend fun getAllMovies(): List<Movie>
+    suspend fun getAllMoviesSortedByTitle(): List<Movie>
 
     @Query("SELECT * FROM movies_table WHERE id = :movieId")
     suspend fun getMovieById(movieId: String?): Movie
