+++ Week8/MovieViewer/app/src/main/java/com/wmw/movieviewer/model/MovieDatabase.kt
@@ -3,7 +3,7 @@ package com.wmw.movieviewer.model
 import androidx.room.Database
 import androidx.room.RoomDatabase
 import androidx.room.TypeConverters
-import com.wmw.movieviewer.data.networking.Converters
+import com.wmw.movieviewer.networking.Converters
 
 @Database(entities = [(Movie::class)], version = 2)
 @TypeConverters(Converters::class)
