+++ Week8/MovieViewer/app/src/main/java/com/wmw/movieviewer/repository/MovieRepository.kt
@@ -4,19 +4,21 @@ import com.wmw.movieviewer.App
 import com.wmw.movieviewer.model.Movie
 
 class MovieRepository {
+//    private var allLiveMovies: LiveData<List<Movie>> = dao.getAllMoviesSortedByTitle()
 
     suspend fun deleteMovieById(movieId: String) = dao.deleteMovieById(movieId)
 
-    suspend fun getAllMovies(): List<Movie> = dao.getAllMovies()
+    suspend fun getAllMovies(): List<Movie> = dao.getAllMoviesSortedByTitle()
 
     suspend fun getMovieById(movieId: String?) = dao.getMovieById(movieId)
 
-    suspend fun storeMoviesIfNotEmpty(movies: List<Movie>) {
-        if (getAllMovies().count() < 1) saveMovies(movies)
-    }
+//    suspend fun storeMoviesIfNotEmpty(movies: List<Movie>) {
+//        val movieList = getAllMovies().value
+//        if (movieList != null && movieList.count() < 1) saveMovies(movies)
+//    }
 
-    suspend fun insertMovie(movie: Movie) = dao.insertMovie(movie)
+    suspend fun insertMovie(movie: Movie) = dao.insert(movie)
 
     private val dao = App.movieDb.movieDao()
-    private suspend fun saveMovies(movies: List<Movie>) = dao.insertMovies(movies)
+//    private suspend fun saveMovies(movies: List<Movie>) = dao.insert(movies)
 }
