+++ Week8/MovieViewer/app/src/main/java/com/wmw/movieviewer/ui/DetailActivity.kt
@@ -1,4 +1,4 @@
-package com.wmw.movieviewer.activities
+package com.wmw.movieviewer.ui
 
 import android.content.Context
 import android.content.Intent
