+++ Week8/MovieViewer/app/src/main/res/layout/activity_login.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".activities.LoginActivity">
+    tools:context=".ui.LoginActivity">
 
     <EditText
         android:id="@+id/usernameEditText"
